修复保存角色权限关系bug

+++ src/main/java/com/xulizhi/demo/service/impl/RoleServiceImpl.java
@@ -124,7 +124,9 @@ public class RoleServiceImpl implements RoleService {
             roleMenu.setRoleId(roleDTO.getId());
             roleMenuService.saveRoleMenuRelation(roleMenu);
         }
-        roleMenuService.deleteRoleMenuRelationByRoleIdAndMenuIds(roleDTO.getId(),oldMenuIds);
+        if(!oldMenuIds.isEmpty()) {
+            roleMenuService.deleteRoleMenuRelationByRoleIdAndMenuIds(roleDTO.getId(), oldMenuIds);
+        }
     }
 
     @Override
